[webservices] [TODO] Webservices, initial main file split, multiple json usage functionality
