+++ workspace/inputData.go
@@ -1,0 +1,9 @@
+package main
+
+type inputUserData struct {
+	FirstName    string `validate:"required"`
+	LastName     string `validate:"required"`
+	Age          int    `validate:"required,gte=18,lte=80"`
+	City         string `validate:"required"`
+	Organization string `validate:"required,contains=/"`
+}
