+++ workspace/outputData.go
@@ -1,0 +1,11 @@
+package main
+
+type outputUserData struct {
+	FullName      string
+	Age           int
+	City          string
+	Organization  string
+	Department    string
+	Subdepartment string
+	Team          string
+}
